+++ CHANGELOG.md
@@ -1,10 +1,42 @@
-## [1.1.15] - 2026-03-01
+## [1.2.0] - 2026-03-08
+
+### 🚀 Features
+
+- Enhanced performance via single find call for tasks
+
 
 ### 🐛 Bug Fixes
 
 - User level tasks are not handled properly #116
 
 - Element with id group:<group_name>:<file>.json is already registered #120
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Updated how the icons are rendered when it is a workspace-task iconUri
+
+- When a user task from a user's profile is loaded, it now shows correctly in the tree
+
+
+### 💼 Other
+
+- V1.1.15 (#124)
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Add file indexing wait in taskFilesService integration tests
+
+- Add file indexing wait in taskFilesService integration tests
 
 
 ### ⚙️ Miscellaneous Tasks
@@ -27,9 +59,21 @@
 
 - Add codecov files to vscodeignore
 
+- Update changelog
+
+- Implement suggestions from copilot.
+
+- Update version info
+
+- Update changelog
+
+- Markdown-lint fix
+
+- Revert eslint version
 
 
-**Full Changelog**: https://github.com/camalot/vscode-workspace-tasks/compare/v1.1.14...v1.1.15
+
+**Full Changelog**: https://github.com/camalot/vscode-workspace-tasks/compare/v1.1.14...v1.2.0
 
 ## [1.1.14] - 2026-02-19